requirements.txt
flask
python-dotenv # for getting the debugging mode on
psycopg2-binary